added alarm in work for reoccurrence
