+++ app/src/main/java/com/example/tidy/Utils.kt
@@ -231,4 +231,14 @@ object Utils {
         return task.repeatType != emptyTask.repeatType || task.dueDateAndTime != emptyTask.dueDateAndTime || task.priority != emptyTask.priority
     }
 
+    fun getNextMidNightMilli(): Long {
+        val nextMidnight = Calendar.getInstance().apply {
+            add(Calendar.DAY_OF_YEAR, 1)
+            set(Calendar.HOUR_OF_DAY, 0)
+            set(Calendar.MINUTE, 0)
+            set(Calendar.SECOND, 0)
+            set(Calendar.MILLISECOND, 0)
+        }
+        return nextMidnight.timeInMillis
+    }
 }
